Implement frequency spreading for generated test signals.
git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/trunk@3020 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ fivehz.f90
@@ -307,7 +307,7 @@ subroutine addnoise(n)
 
   if(txsnrdb.gt.40.0) return
   if(txsnrdb.ne.txsnrdb0) then
-     snr=10.0**(0.05*(txsnrdb-1))
+     snr=10.0**(0.05*(nint(txsnrdb)-1))
      fac=3000.0
      if(snr.gt.1.0) fac=3000.0/snr
      txsnrdb0=txsnrdb
